separate datacard logic into a separate file
