+++ roc_picker/systematics_mc.py
@@ -1,4 +1,4 @@
-import abc, argparse, collections, functools, itertools, matplotlib.pyplot as plt, numbers, numpy as np, pathlib, re, scipy
+import abc, collections, functools, matplotlib.pyplot as plt, numbers, numpy as np, scipy.special
 
 class DistributionBase(abc.ABC):
   @abc.abstractmethod
@@ -280,154 +280,3 @@ class ROCCollection:
       plt.show()
     else:
       plt.savefig(saveas)
-
-def parse_datacard(file_path, id_start=0):
-  """
-  This function was generated using ChatGPT
-  """
-  with open(file_path, 'r') as file:
-    lines = file.readlines()
-
-  data = {
-    "patients": [],
-    "systematics": [],
-    "observable_type": None
-  }
-
-  for line in lines:
-    line = line.strip()
-    if not line or line.startswith('#') or line.startswith('---'):
-      continue
-
-    if line.startswith("observable_type"):
-      data["observable_type"] = line.split()[1]
-    elif line.startswith("bin"):
-      continue
-    elif line.startswith("response"):
-      responses = line.split()[1:]
-      continue
-    elif line.startswith("observable"):
-      if data["observable_type"] != "fixed":
-        raise ValueError("Unexpected 'observable' line for observable_type '{}'".format(data["observable_type"]))
-      values = list(map(float, line.split()[1:]))
-      try:
-        for response, value in zip(responses, values, strict=True):
-          data["patients"].append({
-            "response": response,
-            "value": value
-          })
-      except ValueError as e:
-        raise ValueError("Mismatched lengths in responses and values") from e
-      continue
-    elif line.startswith("count"):
-      if data["observable_type"] != "poisson":
-        raise ValueError("Unexpected 'count' line for observable_type '{}'".format(data["observable_type"]))
-      values = list(map(int, line.split()[1:]))
-      try:
-        for response, value in zip(responses, values, strict=True):
-          data["patients"].append({
-            "response": response,
-            "value": value
-          })
-      except ValueError as e:
-        raise ValueError("Mismatched lengths in responses and values") from e
-      continue
-    elif line.startswith("num"):
-      if data["observable_type"] != "poisson_ratio":
-        raise ValueError("Unexpected 'num' line for observable_type '{}'".format(data["observable_type"]))
-      numerators = list(map(int, line.split()[1:]))
-      continue
-    elif line.startswith("denom"):
-      if data["observable_type"] != "poisson_ratio":
-        raise ValueError("Unexpected 'denom' line for observable_type '{}'".format(data["observable_type"]))
-      denominators = list(map(int, line.split()[1:]))
-      try:
-        for response, num, denom in zip(responses, numerators, denominators, strict=True):
-          data["patients"].append({
-            "response": response,
-            "numerator": num,
-            "denominator": denom
-          })
-      except ValueError as e:
-        raise ValueError("Mismatched lengths in responses, numerators, and denominators") from e
-      continue
-    elif re.match(r'.*\s+lnN\s+.*', line):
-      tokens = line.split()
-      data["systematics"].append({
-        "type": tokens[0],
-        "method": tokens[1],
-        "values": [float(x) if x != '-' else None for x in tokens[2:]]
-      })
-      continue
-    else:
-      raise ValueError("Unexpected line format: {}".format(line))
-
-  if data["observable_type"] not in ["fixed", "poisson", "poisson_ratio"]:
-    raise ValueError("Invalid observable_type: {}".format(data["observable_type"]))
-
-  return data
-
-def plot_from_datacard(file_path, saveas=None, *, id_start=0):
-  """
-  This function was generated using ChatGPT
-  """
-  data = parse_datacard(file_path, id_start=id_start)
-
-  id_generator = itertools.count(id_start)
-  patient_distributions = []
-
-  if data["observable_type"] == "fixed":
-    for p in data["patients"]:
-      fixed_value = p["value"]
-      patient_distributions.append({
-        "response": p["response"],
-        "ratio": fixed_value
-      })
-
-  elif data["observable_type"] == "poisson":
-    for p in data["patients"]:
-      count = ScipyDistribution(nominal=p["value"], scipydistribution=scipy.stats.poisson(mu=p["value"]), id=next(id_generator))
-      patient_distributions.append({
-        "response": p["response"],
-        "ratio": count
-      })
-
-  elif data["observable_type"] == "poisson_ratio":
-    for p in data["patients"]:
-      numerator = ScipyDistribution(nominal=p["numerator"], scipydistribution=scipy.stats.poisson(mu=p["numerator"]), id=next(id_generator))
-      denominator = ScipyDistribution(nominal=p["denominator"], scipydistribution=scipy.stats.poisson(mu=p["denominator"]), id=next(id_generator))
-      ratio = numerator / denominator
-      patient_distributions.append({
-        "response": p["response"],
-        "ratio": ratio
-      })
-
-  # Apply log-normal systematics
-  for systematic in data["systematics"]:
-    if systematic["method"] == "lnN":
-      log_norm_factor = ScipyDistribution(
-        nominal=0,
-        scipydistribution=scipy.stats.norm(),
-        id=next(id_generator)
-      )
-      try:
-        for patient, value in zip(patient_distributions, systematic["values"], strict=True):
-          if value is not None:
-            patient["ratio"] *= value ** log_norm_factor
-      except ValueError as e:
-        raise ValueError("Mismatched lengths in patient distributions and systematic values") from e
-
-  responders = [p["ratio"] for p in patient_distributions if p["response"] == "responder"]
-  nonresponders = [p["ratio"] for p in patient_distributions if p["response"] == "non-responder"]
-
-  rd = ROCDistributions(responders=responders, nonresponders=nonresponders, flip_sign=True)
-  rocs = rd.generate(size=10000, random_state=123456)
-  rocs.plot(saveas=saveas)
-
-def plot_from_datacard_argparse():
-  parser = argparse.ArgumentParser(description="Run MC method from a datacard.")
-  parser.add_argument("datacard", type=pathlib.Path, help="Path to the datacard file.")
-  parser.add_argument("output_file", type=pathlib.Path, help="Path to the output file for the plot.")
-
-  args = parser.parse_args()
-  plot_from_datacard(args.datacard, args.output_file)
