+++ setup.py
@@ -12,7 +12,7 @@ setuptools.setup(
   ],
   entry_points = {
     "console_scripts": [
-      "rocpicker_mc=roc_picker.systematics_mc:plot_from_datacard_argparse",
+      "rocpicker_mc=roc_picker.datacard:plot_systematics_mc",
     ],
   }
 )
